Refactor: 이미지 사이즈 상수화
