+++ src/constants/ui/index.ts
@@ -3,3 +3,5 @@ export const MAX_PAGE_NUMBER = 10;
 export const DATA_LENGTH_PER_PAGE = 20;
 
 export const SKELETON_LENGTH = DATA_LENGTH_PER_PAGE;
+
+export const THUMBNAIL_SIZE = 'w500';
